+++ packages/iov-lisk/src/liskconnection.spec.ts
@@ -77,6 +77,30 @@ describe("LiskConnection", () => {
     expect(() => connection.disconnect()).not.toThrow();
   });
 
+  it("can get existing ticker", async () => {
+    const connection = new LiskConnection(base, liskTestnet);
+    const response = await connection.getTicker("LSK" as TokenTicker);
+    expect(response.data.length).toEqual(1);
+    expect(response.data[0].tokenTicker).toEqual("LSK");
+    expect(response.data[0].tokenName).toEqual("Lisk");
+    expect(response.data[0].sigFigs).toEqual(8);
+  });
+
+  it("produces empty result for non-existing ticker", async () => {
+    const connection = new LiskConnection(base, liskTestnet);
+    const response = await connection.getTicker("ETH" as TokenTicker);
+    expect(response.data.length).toEqual(0);
+  });
+
+  it("can get all tickers", async () => {
+    const connection = new LiskConnection(base, liskTestnet);
+    const response = await connection.getAllTickers();
+    expect(response.data.length).toEqual(1);
+    expect(response.data[0].tokenTicker).toEqual("LSK");
+    expect(response.data[0].tokenName).toEqual("Lisk");
+    expect(response.data[0].sigFigs).toEqual(8);
+  });
+
   it("can get chain ID", async () => {
     const connection = await LiskConnection.establish(base);
     const chainId = connection.chainId();